Claude code poprawił #ifdefr dla esp-idf >5.2, bo byłyu błedy kompilacji

+++ 02_Firmware/01_ESP32/HMI_and_bluetooth/main/StepperOptoPowerOFF/StepperOptoPowerOFF.cpp
@@ -295,7 +295,7 @@ void StepperOptoPowerOFF::moveXSteps(int32_t stepAmount)
  * Returns:
  * NONE
  *---------------------------------------------------------------*/
-void StepperOptoPowerOFF::moveTo_xPosition(uint16_t xPosition)
+void StepperOptoPowerOFF::moveTo_xPosition(int32_t xPosition)
 {
 	int32_t stepAmount;
 	if (xPosition < 0)
